Final Folder restructuring, add Layout to higher order component folder and fixed all associated linking problems
