+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import Aux from '../../../hoc/Aux';
+import Aux from '../../../hoc/Aux/Aux';
 import Button from '../../UI/Button/Button';
 
 class OrderSummary extends Component {
+  // This could be a functional component, doesn't have to be a class now that Modal is Class based"
   componentWillUpdate() {
     console.log('[OrderSummary] WillUpdate')
   }
