+++ src/components/UI/Modal/Modal.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import styles from './Modal.module.css';
-import Aux from '../../../hoc/Aux';
+import Aux from '../../../hoc/Aux/Aux';
 import Backdrop from '../Backdrop/Backdrop';
